# mat64: use asm.Dscal in Vector.ScaleVec()