```
% benchstat before.txt after.txt 
name                 old time/op  new time/op  delta
ScaleVec10Inc1       41.7ns ± 1%  10.9ns ± 1%  -73.94%  (p=0.029 n=4+4)
ScaleVec100Inc1       126ns ± 1%    25ns ± 0%  -79.86%  (p=0.029 n=4+4)
ScaleVec1000Inc1      815ns ± 0%   166ns ± 0%  -79.60%  (p=0.029 n=4+4)
ScaleVec10000Inc1    10.0µs ± 0%   3.0µs ± 0%  -70.29%  (p=0.029 n=4+4)
ScaleVec100000Inc1    128µs ± 0%    52µs ± 0%  -59.69%  (p=0.029 n=4+4)
ScaleVec10Inc2       45.3ns ± 0%  16.6ns ± 0%  -63.38%  (p=0.029 n=4+4)
ScaleVec100Inc2       201ns ± 0%   106ns ± 0%  -47.33%  (p=0.029 n=4+4)
ScaleVec1000Inc2     1.59µs ± 0%  1.01µs ± 0%  -36.25%  (p=0.029 n=4+4)
ScaleVec10000Inc2    15.6µs ± 0%  10.1µs ± 0%  -35.42%  (p=0.029 n=4+4)
ScaleVec100000Inc2    182µs ± 0%   109µs ± 0%  -40.21%  (p=0.029 n=4+4)
ScaleVec10Inc20      45.3ns ± 0%  16.8ns ± 0%  -63.04%  (p=0.029 n=4+4)
ScaleVec100Inc20      202ns ± 1%   106ns ± 0%  -47.46%  (p=0.029 n=4+4)
ScaleVec1000Inc20    1.64µs ± 0%  1.02µs ± 0%  -37.81%  (p=0.029 n=4+4)
ScaleVec10000Inc20   29.1µs ± 0%  22.0µs ± 0%  -24.46%  (p=0.029 n=4+4)
ScaleVec100000Inc20   412µs ± 8%   279µs ± 2%  -32.17%  (p=0.029 n=4+4)
```